design: 메인(홈)-학습, 사전 탭 WordItem 디자인 및 레이아웃 변경
